Implement basic evaluation and "hope chess" search (#2)
This adds a basic material-count algorithm for evaluation and also adds
a "hope chess" search. Basically, it will search its own moves only (and
begrudgingly allow the opponent to move only to get out of checks). With
this implementation, it finds Scholar's Mate every time if not impeded.

There was a fun bug in here in the first pass of implementation. Once it
found mate somewhere in the tree, all other moves were essentially
arbitrary, and it would never checkmate! (Not even when all material was
gone, since everything was equivalent.) I had to tweak its evaluation to
explicitly reward shorter solutions. With that in place, it works as
expected.