Implement transposition tables (#7)
This adds a basic implementation of [transposition
tables](https://www.chessprogramming.org/Transposition_Table) to,
ideally, speed up finding good moves. Unfortunately, it's not doing a
whole lot right now because we're still selecting moves in a "default"
order which is totally unsophisticated, so we're rather unlikely to pick
the best move first (randomly ordering the move list would even be
better than what we have).

This also introduces a basic heuristic for move ordering: shuffling the
list randomly. This turns out to perform better than leaving them in the
original order, because it can move the best move (or a good move, at
least) earlier in the list. The time to select `1. ... e4`, at least, is
much much faster.

This completes #6 .